move this file closer to the only place that uses it
